Fix cycler_wreckage storing time value in int instead of float

+++ dlls/h_cycler.cpp
@@ -391,7 +391,7 @@ class CWreckage : public CBaseMonster
 	void Precache() override;
 	void Think() override;
 
-	int m_flStartTime;
+	float m_flStartTime;
 };
 TYPEDESCRIPTION CWreckage::m_SaveData[] =
 	{
